Update index.pug
Before the bigger changes to the index, I though we should correct the link to the community pages, since there has not bee a "network" for the last 3 years  + the pages are down now. Also changesd the Developer blog to Oskari blog - this can be also something else if needed.